student adding manually

+++ database/migrations/2017_02_03_150148_create_upn_table.php
@@ -17,7 +17,7 @@ class CreateUpnTable extends Migration
             $table->string('upn');
             $table->primary('upn');
             $table->char('check_letter', 1); // This letter, calculated using the digits at characters 2 - 13 verifies a UPN’s validity. 
-            $table->tinyInteger('la_number'); // A 3-digit number, unique to each LA, allocated by DfE.
+            $table->smallInteger('la_number'); // A 3-digit number, unique to each LA, allocated by DfE.
             $table->smallInteger('establishment_number'); // A 4-digit number, unique to each school/academy, allocated by DfE.
             $table->tinyInteger('year_code'); // A 2-digit number, comprising the last 2 numbers of the academic year in which the UPN is allocated.
             $table->tinyInteger('serial_number'); // A 3-digit number, unique to each UPN allocated by a school/academy in a given year.
